refactor(toml): split toml.py such that each filter has its own file.
