+++ .isort.cfg
@@ -4,7 +4,7 @@
 
 [settings]
 known_first_party=ansible_collections.marshallwp.general
-line_length=100
+line_length=180
 lines_after_imports=2
 lines_between_types=1
 profile=black
