+++ .pre-commit-config.yaml
@@ -1,40 +1,46 @@
-# SPDX-FileCopyrightText: 2024 Red Hat, Inc.
+# SPDX-FileCopyrightText: 2025 Industrial Info Resources, Inc.
+# SPDX-FileContributor: William P. Marshall
 #
-# SPDX-License-Identifier: Apache-2.0
+# SPDX-License-Identifier: GPL-3.0-or-later
 
 ---
 repos:
-  - repo: https://github.com/ansible-network/collection_prep
-    rev: 1.1.1
-    hooks:
-      - id: update-docs
+  # - repo: https://github.com/ansible-network/collection_prep
+  #   rev: 1.1.2
+  #   hooks:
+  #     - id: update-docs
 
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.4.0
+    rev: v6.0.0
     hooks:
+      - id: check-case-conflict
+      - id: check-illegal-windows-names
       - id: check-merge-conflict
       - id: check-symlinks
       - id: debug-statements
       - id: end-of-file-fixer
+      - id: fix-byte-order-marker
+      - id: mixed-line-ending
+        args: ['--fix=lf']
       - id: no-commit-to-branch
         args: [--branch, main]
       - id: trailing-whitespace
 
   - repo: https://github.com/asottile/add-trailing-comma
-    rev: v3.0.0
+    rev: v4.0.0
     hooks:
       - id: add-trailing-comma
 
-  - repo: https://github.com/pre-commit/mirrors-prettier
-    rev: "v3.0.0"
-    hooks:
-      - id: prettier
-        entry: env CI=1 bash -c "prettier --list-different . || ec=$? && prettier --loglevel=error --write . && exit $ec"
-        pass_filenames: false
-        args: []
-        additional_dependencies:
-          - prettier
-          - prettier-plugin-toml
+  # - repo: https://github.com/pre-commit/mirrors-prettier
+  #   rev: "v3.0.0"
+  #   hooks:
+  #     - id: prettier
+  #       entry: env CI=1 bash -c "prettier --list-different . || ec=$? && prettier --loglevel=error --write . && exit $ec"
+  #       pass_filenames: false
+  #       args: []
+  #       additional_dependencies:
+  #         - prettier
+  #         - prettier-plugin-toml
 
   - repo: https://github.com/PyCQA/isort
     rev: 5.12.0
@@ -44,14 +50,14 @@ repos:
         args: ["--filter-files"]
 
   - repo: https://github.com/psf/black
-    rev: 23.7.0
+    rev: 25.11.0
     hooks:
       - id: black
 
-  - repo: https://github.com/pycqa/flake8
-    rev: 7.0.0
-    hooks:
-      - id: flake8
+  # - repo: https://github.com/pycqa/flake8
+  #   rev: 7.0.0
+  #   hooks:
+  #     - id: flake8
 
 # SPDX-SnippetBegin
 # SPDX-SnippetCopyrightText: 2017 Free Software Foundation Europe e.V. <https://fsfe.org>
